requirements.txt
langchain==0.1.0
langchain-huggingface==0.0.1
langchain-community==0.4.1
langchain-core==1.0.4
langchain-text-splitters==1.0.0
huggingface_hub==0.20.1
transformers==4.36.2
torch==2.1.2
streamlit==1.29.0
chromadb==0.4.18
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
sentence-transformers==2.2.2
flask==3.1.2
aiohttp==3.13.2
sqlalchemy==2.0.44
pydantic==2.12.4
pydantic-settings==2.12.0
requests==2.32.5
tenacity==9.1.2
